+++ tests/3.0.x/__snapshots__/pluginZod/dateTypeDate/zod/orderSchema.ts
@@ -9,7 +9,7 @@ export const orderSchema = z.object({
     "id": z.int().optional(),
     "petId": z.int().optional(),
     "quantity": z.int().optional(),
-    "shipDate": z.iso.datetime().optional(),
+    "shipDate": z.date().optional(),
     "status": z.enum(["placed", "approved", "delivered"]).optional().describe("Order Status"),
     "complete": z.boolean().optional()
     })